refactor: extract shared ContentAwareReasoning from AgentBase and CobolAnalyzerAgent
- Create ContentAwareReasoning static helper with all shared logic:
  complexity scoring regexes, CalculateTokenSettings, ApplyModelSpecificOptions,
  CalculateComplexityScore, DetectTruncation, ExtractCobolSourceFromPrompt,
  EstimateTokens, CompileIndicators, ClampProfileToModel
- AgentBase and CobolAnalyzerAgent delegate to shared class
- Eliminates ~450 lines of duplicated code
- Build: 0 errors, 0 warnings (all 4 projects)
